requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
supabase>=2.0.0
python-dotenv>=1.0.0
feedparser>=6.0.0
